refactorrearrange datalyer file \n BREAKING CHANGE: #1 iteration refcator to separate client from server invocation code
