+++ lib/tests/dataLayer.test.js
@@ -1,14 +1,15 @@
 import mockedAxios from "axios";
+import { crudTemplates } from "../integration/crud";
 import {
-  submitToAPI,
-  extractFormData,
-  rehydrateFormResponses,
   buildFormDataObject,
-  crudTemplates,
   getFormByID,
-  getSubmissionByID,
   getFormByStatus,
-} from "../dataLayer";
+  getSubmissionByID,
+  rehydrateFormResponses,
+  extractFormData,
+  submitToAPI,
+} from "../integration/common";
+
 import { logMessage } from "../logger";
 import TestForm from "./testData";
