+++ lib/tests/markdown.test.js
@@ -1,6 +1,6 @@
 import markdown from "../markdown";
 
-jest.mock("../dataLayer.tsx", () => ({
+jest.mock("../integration/common", () => ({
   extractFormData: jest.fn(() => ["one", "two", "three"]),
   rehydrateFormResponses: jest.fn(() => null),
 }));
