+++ lib/tsUtils.tsx
@@ -4,3 +4,7 @@ export function hasOwnProperty<X extends Record<string, unknown>, Y extends Prop
 ): obj is X & Record<Y, unknown> {
   return Object.prototype.hasOwnProperty.call(obj, prop);
 }
+
+export const isServer = (): boolean => {
+  return typeof window === "undefined";
+};
