+++ pages/id/[form]/[[...step]].js
@@ -1,4 +1,4 @@
-import { getFormByID } from "../../../lib/dataLayer";
+import { getFormByID } from "../../../lib/integration/common";
 import DynamicForm from "../../../components/containers/DynamicForm/DynamicForm";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 import { checkOne } from "../../../lib/flags";
